# Use ProtoStream to encode sketches