### What does this PR do?

This rewrites the sketch_series marhsaling to use ProtoStream (c.f. #9008).

In running the existing benchmarks, this is about 8% faster, but more importantly saves 50% on allocations / RAM.

```
name                          old time/op        new time/op        delta
MarshalSplitCompress1-16            1.97ms ± 0%        1.91ms ± 0%   -2.82%  (p=0.000 n=10+10)
MarshalSplitCompress10-16           18.1ms ± 0%        16.6ms ± 0%   -8.24%  (p=0.000 n=10+10)
MarshalSplitCompress100-16           176ms ± 1%         162ms ± 0%   -8.10%  (p=0.000 n=10+10)
MarshalSplitCompress1000-16          1.74s ± 0%         1.59s ± 0%   -8.37%  (p=0.000 n=9+9)
MarshalSplitCompress10000-16         17.4s ± 0%         15.9s ± 1%   -8.31%  (p=0.016 n=4+5)

name                          old payload-bytes  new payload-bytes  delta
MarshalSplitCompress1-16            5.15kB ± 0%        5.15kB ± 0%     ~     (all equal)
MarshalSplitCompress10-16           50.0kB ± 0%        50.0kB ± 0%     ~     (all equal)
MarshalSplitCompress100-16           498kB ± 0%         498kB ± 0%     ~     (all equal)
MarshalSplitCompress1000-16         4.98MB ± 0%        4.98MB ± 0%     ~     (all equal)
MarshalSplitCompress10000-16        49.8MB ± 0%        49.8MB ± 0%     ~     (all equal)

name                          old payloads       new payloads       delta
MarshalSplitCompress1-16              1.00 ± 0%          1.00 ± 0%     ~     (all equal)
MarshalSplitCompress10-16             1.00 ± 0%          1.00 ± 0%     ~     (all equal)
MarshalSplitCompress100-16            2.00 ± 0%          2.00 ± 0%     ~     (all equal)
MarshalSplitCompress1000-16           15.0 ± 0%          15.0 ± 0%     ~     (all equal)
MarshalSplitCompress10000-16           145 ± 0%           145 ± 0%     ~     (all equal)

name                          old alloc/op       new alloc/op       delta
MarshalSplitCompress1-16            1.41MB ± 0%        1.32MB ± 0%   -6.74%  (p=0.000 n=9+9)
MarshalSplitCompress10-16           6.93MB ± 0%        4.37MB ± 0%  -36.96%  (p=0.000 n=9+9)
MarshalSplitCompress100-16          52.8MB ± 0%        25.6MB ± 0%  -51.53%  (p=0.000 n=9+10)
MarshalSplitCompress1000-16          445MB ± 0%         172MB ± 0%  -61.48%  (p=0.000 n=10+8)
MarshalSplitCompress10000-16        4.37GB ± 0%        1.63GB ± 0%  -62.69%  (p=0.016 n=4+5)

name                          old allocs/op      new allocs/op      delta
MarshalSplitCompress1-16               857 ± 0%           469 ± 0%  -45.27%  (p=0.000 n=10+10)
MarshalSplitCompress10-16            8.21k ± 0%         4.15k ± 0%  -49.47%  (p=0.000 n=10+10)
MarshalSplitCompress100-16           81.7k ± 0%         40.9k ± 0%  -49.93%  (p=0.000 n=10+10)
MarshalSplitCompress1000-16           816k ± 0%          408k ± 0%  -49.97%  (p=0.000 n=10+8)
MarshalSplitCompress10000-16         8.16M ± 0%         4.08M ± 0%  -49.98%  (p=0.029 n=4+4)
```

### Motivation

The existing code did a bit of this manual assembly of PB messages, but with
```
               // Magic number that occurs before the varint encoding
```

ProtoStream regularizes that a bit.

### Describe how to test your changes

* Configure an agent to produce large quantities of sketches (so that multiple payloads will be required)
* Observe no crashes, see the sketches in the app

### Checklist

- [x] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [x] The appropriate `team/..` label has been applied, if known.
- [x] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.