@djmitche 2021-09-14
This has been updated to use the ProtoStream functionality that just landed in Molecule.  This is a bit more efficient, so we see better CPU numbers than the earlier draft (5% improvement to 8.5% improvement).  The payloads are identical.  And the memory usage is *substantially* down.