@sexcom-7 2021-09-14
Todo tasks

On Tue, Sep 14, 2021 at 3:50 PM Brian Floersch ***@***.***>
wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In pkg/metrics/sketch_series.go
> <https://github.com/DataDog/datadog-agent/pull/9009#discussion_r708622326>
> :
>
> > +	const distributionSum = 6
> +	const distributionV = 7
> +	const distributionG = 8
> +	const distributionDelta = 9
> +	const distributionBuf = 10
> +	*/
> +	const dogsketchTs = 1
> +	const dogsketchCnt = 2
> +	const dogsketchMin = 3
> +	const dogsketchMax = 4
> +	const dogsketchAvg = 5
> +	const dogsketchSum = 6
> +	const dogsketchK = 7
> +	const dogsketchN = 8
> +
> +	// generate a footer containing an empty Metadata field (TODO: this isn't
>
> Is this TODO addressable now? or is this a future task
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/DataDog/datadog-agent/pull/9009#pullrequestreview-754412016>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ATTQVRFQ5WNRUPZSIZUY7Z3UB6YR7ANCNFSM5C56TWYQ>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>