@gh123man 2021-09-14
I think `encodeVarintAgentPayload` can be removed now