@djmitche 2021-09-15
Those are both good points, and I hope to revisit them when looking more closely at how the compression part of serialization is implemented.