@djmitche 2021-09-17
In the benchmark environment, handling 160,000 distinct sketches, serialization took about 2.43s per 15s.  The vast majority of that was compression

![image](https://user-images.githubusercontent.com/28673/133836815-edf3caac-b1eb-41a0-b71c-88dca04ec4da.png)

Ignoring compression leaves about 250ms per 15s:

![image](https://user-images.githubusercontent.com/28673/133837196-9a238e07-093e-4dec-abe9-74214f228bb2.png)

80ms of that is in the call to `.Cols()`, which would have to be done anyway.  So overall the actual PB serialization is trivial here.  A 5% improvement is likely to work out to 10ms per 15s only in this highly-loaded environment, so I doubt we'll see any noticeable change in CPU usage when this PR is deployed.